refactor : 날짜 변환 기능 store 담당당
