+++ src/components/todo/TodoItem.tsx
@@ -1,5 +1,6 @@
 import { useState, useEffect, useRef } from 'react';
 import { Todo } from '@/types/todo';
+import { formatDate } from '@/store/todoStore';
 
 interface TodoItemProps {
   todo: Todo;
@@ -7,15 +8,6 @@ interface TodoItemProps {
   onMenuToggle: () => void;
   onMenuClose: () => void;
 }
-
-// 날짜를 yyyy-MM-dd 형식으로 변환하는 헬퍼 함수
-const formatDate = (dateString: string): string => {
-  const date = new Date(dateString);
-  const year = date.getFullYear();
-  const month = ('0' + (date.getMonth() + 1)).slice(-2);
-  const day = ('0' + date.getDate()).slice(-2);
-  return `${year}-${month}-${day}`;
-};
 
 export default function TodoItem({
   todo,
